requirements.txt
nltk
scikit-learn
numpy
torch
opencv-python~=4.12.0.88
matplotlib~=3.10.7
pandas~=2.3.3
seaborn~=0.13.2
pillow~=12.0.0
ultralytics~=8.3.228